vint64: factor modules into files
Separates the `signed` and `error` modules into their own files, and
moves the `zigzag` module under `signed`.
